pyproject.toml
[tool.poetry]
name = "dnd"
version = "0.1.0"
description = ""
authors = ["David Morrison <drmorr@evokewonder.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
bokeh = "^3.0.2"
pytest = "^7.2.0"
mock = "^4.0.3"
selenium = "^4.6.0"
pre-commit = "^2.20.0"
flake8 = "^5.0.4"
autopep8 = "^2.0.0"

[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
types-pyyaml = "^6.0.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
